+++ slim/__init__.py
@@ -7,4 +7,4 @@ from . import base
 from . import support
 from . import utils
 
-__version__ = '0.5.3'
+__version__ = '0.5.4'