+++ slim/base/app.py
@@ -149,7 +149,10 @@ class Application:
         if cors:
             for r in list(self._raw_app.router.routes()):
                 if type(r.resource) != StaticResource and r.handler not in ws_set:
-                    cors.add(r)
+                    try:
+                        cors.add(r)
+                    except ValueError:
+                        pass
 
         for vi in self.route.views:
             vi.view_cls._ready()