+++ tests/base_tests/test_sqlquery_info.py
@@ -148,15 +148,8 @@ async def test_condition_bind_error_convert_failed():
 
 
 async def test_condition_bind_error_in_or_not_in_value():
-    # TODO: add condition 时就根据 in 这种运算符做了特殊解析，认为值校验的时机不一致，需要重新考量
-    pass
-    '''
     sqi = SQLQueryInfo()
-    sqi.parse_then_add_condition('name', 'in', 1)
-    view: PeeweeView = await make_mocked_view_instance(app, ATestView, 'GET', '/api/test1')
 
     with pytest.raises(InvalidParams) as e:
-        sqi.bind(view)
-
-    assert 'name' in e.value.args[0]
-    '''
+        sqi.parse_then_add_condition('name', 'in', [1, 2])
+        assert 'name' in e.value.args[0]